requirements.txt
customtkinter==5.2.2
pymupdf==1.23.26
Pillow==10.2.0